refactor(googlemeet): address review feedback
