+++ src/providers/google_analytics/executors.ts
@@ -13,7 +13,7 @@ import {
   optionalStringOrNull,
   requiredRecord,
 } from "../../core/cast.ts";
-import { googleJsonRequest } from "../googledrive/runtime-shared.ts";
+import { googleJsonRequest } from "../google-runtime.ts";
 import { defineOAuthProviderExecutors, defineProviderProxy, ProviderRequestError } from "../provider-runtime.ts";
 
 const service = "google_analytics";
